288: Minor updates
