+++ backend/src/report/mod.rs
@@ -1,2 +1,3 @@
 pub mod report;
 pub mod report_error;
+pub mod report_status;
